Fix migration (2.1.2 migration) that causes installations to fail (due to plugin settings not existing for new projects at time of execution)

+++ updates/clear_old_cache_dir.php
@@ -3,24 +3,26 @@
 namespace ABWebDevelopers\Portfolio\Updates;
 
 use ABWebDevelopers\ImageResize\Classes\Resizer;
-use Schema;
-use October\Rain\Database\Updates\Migration;
 use ABWebDevelopers\ImageResize\Models\Settings;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\File;
-use Illuminate\Support\Facades\Storage;
+use October\Rain\Database\Updates\Migration;
 
 class ClearOldCacheDir extends Migration
 {
     public function up()
     {
-        $oldpath = base_path('storage/app/media/imageresizecache');
+        // Check that the settings for this plugin exist
+        if (DB::table('system_settings')->where('item', 'abwebdevelopers_imageresize')->exists()) {
+            $oldpath = base_path('storage/app/media/imageresizecache');
 
-        if (Settings::instance()->cache_directory !== $oldpath) {
-            if (is_dir($oldpath)) {
-                Resizer::clearFiles(null, $oldpath);
+            if (Settings::instance()->cache_directory !== $oldpath) {
+                if (is_dir($oldpath)) {
+                    Resizer::clearFiles(null, $oldpath);
 
-                File::deleteDirectory($oldpath);
-                @unlink($oldpath);
+                    File::deleteDirectory($oldpath);
+                    @unlink($oldpath);
+                }
             }
         }
     }
